move src/mlmd/generated to src/third_party/mlmd. remove protos
