+++ frontend/.prettierignore
@@ -1,2 +1,2 @@
 src/generated
-src/mlmd/generated
+src/third_party
