--- frontend/proto/ml_metadata/README.md
@@ -1,5 +1,0 @@
-This directory contains protos that were manually downloaded from:
-
-[https://github.com/google/ml-metadata/tree/master/ml_metadata/proto](https://github.com/google/ml-metadata/tree/master/ml_metadata/proto)
-
-TODO: What version of the protos is the server using and how can we stay in sync?
