+++ frontend/scripts/gen_grpc_web_protos.js
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2021 The Kubeflow Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 // NOTE: This script is expected to run from root directory.
 // But the frontend/package.json has handled the path navigation for you.
 // Make sure you are under path <parent_directory>/pipelines/frontend.
@@ -7,7 +23,7 @@ const path = require('path');
 const {spawn} = require('child_process');
 
 // TODO: Build process should remove the existing generated proto definitions.
-const OUT_DIR = path.join(__dirname, '..', 'src', 'mlmd', 'generated');
+const OUT_DIR = path.join(__dirname, '..', 'src', 'third_party', 'mlmd', 'generated');
 
 if (!fs.existsSync(OUT_DIR)) {
   fs.mkdirSync(OUT_DIR, {
