+++ frontend/src/components/viewers/MetricsVisualizations.tsx
@@ -14,13 +14,14 @@
  * limitations under the License.
  */
 
-import { Artifact, ArtifactType, getMetadataValue } from '../../mlmd/library';
 import HelpIcon from '@material-ui/icons/Help';
 import React from 'react';
-import { Array as ArrayRunType, Number, Failure, Record, String, ValidationError } from 'runtypes';
+import { Array as ArrayRunType, Failure, Number, Record, String, ValidationError } from 'runtypes';
 import IconWithTooltip from 'src/atoms/IconWithTooltip';
 import { color, padding } from 'src/Css';
 import { filterArtifactsByType } from 'src/lib/MlmdUtils';
+import { getMetadataValue } from 'src/mlmd/library';
+import { Artifact, ArtifactType } from 'src/third_party/mlmd';
 import Banner from '../Banner';
 import ConfusionMatrix, { ConfusionMatrixConfig } from './ConfusionMatrix';
 import PagedTable from './PagedTable';
