+++ frontend/src/lib/MlmdUtils.test.ts
@@ -19,7 +19,7 @@ import {
   Context,
   GetContextByTypeAndNameRequest,
   GetContextByTypeAndNameResponse,
-} from '../mlmd/library';
+} from 'src/mlmd/library';
 import { expectWarnings, testBestPractices } from 'src/TestUtils';
 import { Workflow, WorkflowSpec, WorkflowStatus } from 'third_party/argo-ui/argo_template';
 import { getRunContext } from './MlmdUtils';
