+++ frontend/src/pages/ArtifactDetails.tsx
@@ -16,15 +16,18 @@
 
 import {
   Api,
-  Artifact,
   ArtifactCustomProperties,
   ArtifactProperties,
-  GetArtifactsByIDRequest,
   getResourceProperty,
   LineageResource,
   LineageView,
+} from 'src/mlmd/library';
+import {
   ArtifactType,
-} from '../mlmd/library';
+  Artifact,
+  GetArtifactsByIDRequest,
+  GetArtifactTypesByIDRequest,
+} from 'src/third_party/mlmd';
 import { CircularProgress } from '@material-ui/core';
 import * as React from 'react';
 import { Route, Switch } from 'react-router-dom';
@@ -36,7 +39,6 @@ import { ToolbarProps } from '../components/Toolbar';
 import { commonCss, padding } from '../Css';
 import { logger, serviceErrorToString, titleCase } from '../lib/Utils';
 import { Page, PageProps } from './Page';
-import { GetArtifactTypesByIDRequest } from '../mlmd/generated/ml_metadata/proto/metadata_store_service_pb';
 
 export enum ArtifactDetailsTab {
   OVERVIEW = 0,
