+++ frontend/src/pages/RunDetails.tsx
@@ -14,7 +14,6 @@
  * limitations under the License.
  */
 
-import { Context, Execution } from '../mlmd/library';
 import CircularProgress from '@material-ui/core/CircularProgress';
 import InfoIcon from '@material-ui/icons/InfoOutlined';
 import { flatten } from 'lodash';
@@ -27,6 +26,7 @@ import { GkeMetadata, GkeMetadataContext } from 'src/lib/GkeMetadata';
 import { useNamespaceChangeEvent } from 'src/lib/KubeflowClient';
 import { ExecutionHelpers, getExecutionsFromContext, getRunContext } from 'src/lib/MlmdUtils';
 import { isV2Pipeline } from 'src/lib/v2/WorkflowUtils';
+import { Context, Execution } from 'src/third_party/mlmd';
 import { classes, stylesheet } from 'typestyle';
 import {
   NodePhase as ArgoNodePhase,
